% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop_sim.R
\name{pop_sim}
\alias{pop_sim}
\title{Simulate summer Chinook population}
\usage{
pop_sim(
  n_years = 25,
  start_year = 2017,
  n_iter = 500,
  init_S = internal_data$init_S,
  MREER_matrix = internal_data$MREER_out[, , ],
  age_prop_array = internal_data$age_props[, , , ],
  SR_err = internal_data$SR_err[, , ],
  pHOS_err = internal_data$pHOS_err[, , ],
  NOB_err = internal_data$NOB_err[, , ],
  pfmc_err = internal_data$pfmc_err[, ],
  in_river_err = internal_data$in_river_err[, , ],
  smolts = hatchery_smolt_fun(smolts_err = internal_data$smolt_err),
  hatchery_mark_rate = internal_data$hatch_MR_mu,
  HO_broodstock_need = 2000,
  NO_broodstock_target = c(Methow = 122, Okanogan = 650, Wenatchee = 310),
  ...
)
}
\arguments{
\item{n_years}{integer number of return years to simulate forward}

\item{start_year}{first brood year of simulation}

\item{init_S}{4 x 6 matrix with the spawner (or smolt) abundance of each population in the first six years}

\item{MREER_matrix}{age x brood year matrix of simulated mature run exploitation rates}

\item{age_prop_array}{age x population x brood year array of age proportions by population and brood year}

\item{SR_err}{population x brood year matrix of recruitment deviations}

\item{pHOS_err}{population x return year matrix of pHOS errors}

\item{NOB_err}{population x return year matrix of natural origin broodstock collection errors. Exponentiated!}

\item{pfmc_err}{vector of return year annual deviations from PFMC AEQ ocean mort abundance}

\item{in_river_err}{2 x year matrix of return year errors in in-river harvest for treaty and non-treaty}

\item{smolts}{vector of brood year annual smolt releases}

\item{...}{}
}
\description{
Simulate summer Chinook population
}
