% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hatchery_functions.R
\name{HOS_fun}
\alias{HOS_fun}
\title{Hatchery origin spawners. Assumes available.}
\usage{
HOS_fun(model, HOE, HOS_mod_coefs = internal_data$HOS_mod_coefs, HOS_err)
}
\arguments{
\item{model}{HOS model. options are
\itemize{
\item "zero" no hatchery spawning
\item "HOE" hatchery-origin spawners are a function of hatchery escapement, fit to data from  2010-2022
}}

\item{HOE}{hatchery origin escapement. only need if using model option ""}

\item{HOS_mod_coefs}{coefficients of the pHOS model: intercepts, effect of hatchery escapement only need if using model option ""}

\item{HOS_err}{annual deviations}
}
\value{
vector of three reals (hatchery origin spawners)
}
\description{
Hatchery origin spawners. Assumes available.
}
